The parent configuration for the Zurich implementation is moved to scr/zurich_parcel/configs/zurich_parcel.xml.
A child config is inserted instead. The child config is intended to be an individual testing config. Approved changes shall than be moved to the parent config. This helps to avoid conflicts when commiting project configs. For example when different import paths are present or different orders of models appear.

git-svn-id: https://svn.urbansim.org/project_configs@17834 1513fa5d-83ce-422e-ae47-6967bb0fb594