bayarea: fix travel model base directory
This is a quick fix.  When I inherit the No_Project scenario and
override some of the travel model settings, some of them work as
expected (e.g., models, data_exchange_dir).  But for some reason
the travel_model_base_directory takes this dummy value from the
travel model base configuration.  Not sure why.  This patch is a
work-around.

git-svn-id: https://svn.urbansim.org/project_configs@18665 1513fa5d-83ce-422e-ae47-6967bb0fb594